ui: add starting page, routing buttons, and placeholders for login and register
