+++ app/authentication/login/page.tsx
@@ -1,0 +1,9 @@
+/** @format */
+
+export default function Login() {
+    return (
+        <div>
+            <h1>Login</h1>
+        </div>
+    )
+}
