+++ app/authentication/register/page.tsx
@@ -1,0 +1,9 @@
+/** @format */
+
+export default function Register() {
+    return (
+        <div>
+            <h1>Register</h1>
+        </div>
+    )
+}
