+++ app/layout.tsx
@@ -26,9 +26,9 @@ export default function RootLayout({
     children: React.ReactNode
 }>) {
     return (
-        <html lang='en'>
+        <html lang='en' suppressHydrationWarning className='h-full'>
             <body
-                className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+                className={`${geistSans.variable} ${geistMono.variable} antialiased h-full flex flex-col justify-center content-center`}
             >
                 {children}
             </body>
